pyproject.toml
[tool.poetry]
authors = ["Allen Woods <allen.woods@outlook.com>"]
description = "LLM Alignment and Safety Toolkit."
documentation = "https://last.github.io"
homepage = "https://github.com/last-lab/last"
keywords = ["LLM", "Alignment", "Safety", "Toolkit"]
license = "Apache-2.0"
name = "last"
packages = [
    { include = "last" },
]
readme = "README.md"
repository = "https://github.com/last-lab/last.git"
version = "0.0.1"

[tool.poetry.dependencies]
Babel = "*"
aiobotocore = { version = "*", optional = true }
aiofiles = "*"
redis = "^4.2.0rc1"
asyncoss = { version = "*", optional = true }
bcrypt = "*"
captcha = "*"
fastapi = { version = "*", extras = ["all"] }
jinja2 = "*"
python = ">=3.8.1,<4.0"
python-multipart = "*"
tortoise-orm = "*"
httpx = "*"
pendulum = "*"
requests = "*"
pycryptodome = "*"
pydantic = "*"
langchain = ">=0.0.310" 
openai = "*"
aiohttp = "*"
pandas = "*"
openpyxl = "*"

[tool.poetry.dev-dependencies]
# test
pytest = "*"
pytest-asyncio = "*"
pytest-mock = "*"
pytest-xdist = "*"
pytest-cover = "*"
coverage = "*"
# lint
black = "*"
flake8 = "*"
isort = "*"
mypy = "*"
pylint = "*"
# example
asyncmy = "*"
python-dotenv = "*"
rearq = { git = "https://github.com/long2ice/rearq.git", branch = "dev" }
asyncpg = "*"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "setuptools"]

[tool.poetry.extras]
oss = ["asyncoss"]
s3 = ["aiobotocore"]

[tool.coverage.paths]
source = ["tests"]

[tool.coverage.report]
show_missing = true
